refactor(mutations): split CSV cell parsing out to hold the size budget
The skip-instead-of-fabricate fix took packages/mutations/src/csv-connector.ts
to 604 lines against a budget of 573, failing `check-module-size`. The budget
is pinned at main's exact size, so the file cannot absorb a net addition and
the gate's own advice is to shrink or split.

Splitting is the honest option here rather than compressing prose: `parseValue`
never referenced `this`, so it was already a free function in a class's
clothing, and the sentinel exists only to be returned by it. Both move to
`csv-parse-value.ts` and are imported back. csv-connector.ts is now 558 lines.

No behaviour change: same switch arms, same sentinel identity (one module-level
Symbol, imported rather than redeclared), same call site semantics. The Boolean
arm gains braces its `case` needed anyway to scope `lower`. 212 mutations tests
pass, tsc clean, check-module-size OK.
